pyproject.toml
[project]
name = "quickmark"
dynamic = ["version"]
description = "Rust Markdown parsing, with python and CLI interface"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">3.10"
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Rust",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: Markup",
]
dependencies = [
    "maturin[patchelf]>=1.8.3",
    "pytest<=8.1",
    "tox>=4.24.2",
]

[package.metadata.maturin]
name = "quickmark_build"

[build-system]
requires = ["maturin>=1.0,<2.0"]
build-backend = "maturin"


[project.optional-dependencies]
testing = [
    "pytest",
    "pytest-param-files",
]
benchmarking = [
    "pytest",
    "pytest-benchmark",
    "markdown-it-py",
]

[project.scripts]
markdown-it-pyrs = "pymdrs.cli:main"

[tool.maturin]
features = ["pyo3/extension-module"]
python-source = "python"

[tool.uv]
package = true

[tool.isort]
profile = "black"
force_sort_within_sections = true

[tool.ruff]
line-length = 80
extend-select = ["B0", "C4", "ICN", "ISC", "N", "RUF", "SIM"]

[tool.mypy]
show_error_codes = true
strict = true
